Move dedupe code to dedupe.go
